+++ src/hooks.ts
@@ -135,18 +135,39 @@ function subscribeToInitialization(
         });
         return;
       }
-      hooksLogger.info(`Client did not become ready before timeout of ${timeout}ms, reason="${res.reason || ''}"`);
-      onInitStateChange({
-        clientReady: false,
-        didTimeout: true,
-      });
-      res.dataReadyPromise!.then(() => {
-        hooksLogger.info('Client became ready after timeout already elapsed');
-        onInitStateChange({
-          clientReady: true,
-          didTimeout: true,
-        });
-      });
+
+      switch (res.reason) {
+        // Optimizely client failed to initialize.
+        case 'NO_CLIENT':
+          hooksLogger.warn(`Client not ready, reason="${res.message}"`);
+          onInitStateChange({
+            clientReady: false,
+            didTimeout: false,
+          });
+          res.dataReadyPromise!.then(() => {
+            hooksLogger.info('Client became ready.');
+            onInitStateChange({
+              clientReady: true,
+              didTimeout: false,
+            });
+          });
+          break;
+        // Assume timeout for all other cases.
+        // TODO: Other reasons may fall into this case - need to update later to specify 'TIMEOUT' case and general fallback case.
+        default:
+          hooksLogger.info(`Client did not become ready before timeout of ${timeout}ms, reason="${res.message}"`);
+          onInitStateChange({
+            clientReady: false,
+            didTimeout: true,
+          });
+          res.dataReadyPromise!.then(() => {
+            hooksLogger.info('Client became ready after timeout already elapsed');
+            onInitStateChange({
+              clientReady: true,
+              didTimeout: true,
+            });
+          });
+      }
     })
     .catch(() => {
       hooksLogger.error(`Error initializing client. The core client or user promise(s) rejected.`);
@@ -236,7 +257,7 @@ export const useExperiment: UseExperiment = (experimentKey, options = {}, overri
         }));
       });
     }
-    return (): void => {};
+    return (): void => { };
   }, [optimizely.getIsReadyPromiseFulfilled(), options.autoUpdate, optimizely, experimentKey, getCurrentDecision]);
 
   useEffect(
@@ -329,7 +350,7 @@ export const useFeature: UseFeature = (featureKey, options = {}, overrides = {})
         }));
       });
     }
-    return (): void => {};
+    return (): void => { };
   }, [optimizely.getIsReadyPromiseFulfilled(), options.autoUpdate, optimizely, featureKey, getCurrentDecision]);
 
   return [state.isEnabled, state.variables, state.clientReady, state.didTimeout];
@@ -358,11 +379,11 @@ export const useDecision: UseDecision = (flagKey, options = {}, overrides = {}) 
     const decisionState = isClientReady
       ? getCurrentDecision()
       : {
-          decision: createFailedDecision(flagKey, 'Optimizely SDK not configured properly yet.', {
-            id: overrides.overrideUserId || null,
-            attributes: overrideAttrs,
-          }),
-        };
+        decision: createFailedDecision(flagKey, 'Optimizely SDK not configured properly yet.', {
+          id: overrides.overrideUserId || null,
+          attributes: overrideAttrs,
+        }),
+      };
     return {
       ...decisionState,
       clientReady: isClientReady,
@@ -428,7 +449,7 @@ export const useDecision: UseDecision = (flagKey, options = {}, overrides = {}) 
         }));
       });
     }
-    return (): void => {};
+    return (): void => { };
   }, [optimizely.getIsReadyPromiseFulfilled(), options.autoUpdate, optimizely, flagKey, getCurrentDecision]);
 
   return [state.decision, state.clientReady, state.didTimeout];